+++ greenfield/eldenring/data.py
@@ -5531,4 +5531,4 @@ GESTURE_AWARD_FLAGS = {
     60864: (115, 1075750828, 'O Mother'),
 }
 
-_GEN_STAMP = {'inputs_hash': 'sha256:83f166a698f1826a21ffcd02069d69add09aa2bab2d1eb2b10b8adc88fcd2bbe', 'module': 'data.py', 'body_sha256': 'sha256:2d52dbdccb51f88301c070ad1a48b1e83520b3e6831b477e40d3d0cf38f7bdb8'}
+_GEN_STAMP = {'inputs_hash': 'sha256:9101b54b8a7131712af5cf9480f52ca528a65d67d9d8196a931827254789c3bf', 'module': 'data.py', 'body_sha256': 'sha256:2d52dbdccb51f88301c070ad1a48b1e83520b3e6831b477e40d3d0cf38f7bdb8'}